requirements.txt
anyio==3.5.0
appdirs==1.4.4
asgiref==3.5.1
asttokens==2.0.5
backcall==0.2.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
cycler==0.11.0
debugpy==1.6.0
decorator==5.1.1
entrypoints==0.4
executing==0.8.3
fastapi==0.76.0
fonttools==4.33.3
fsspec==2022.3.0
greenlet==1.1.2
h11==0.13.0
humanize==4.1.0
idna==3.3
importlib-metadata==4.11.3
importlib-resources==5.7.1
ipykernel==6.13.0
ipython==8.3.0
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
jupyter-client==7.3.0
jupyter-core==4.10.0
kiwisolver==1.4.2
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
nest-asyncio==1.5.5
numpy==1.22.3
packaging==21.3
pandas==1.4.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.0
pins==0.2.0
prompt-toolkit==3.0.29
psutil==5.9.0
psycopg2==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==1.9.0
Pygments==2.12.0
pyparsing==3.0.8
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
PyYAML==6.0
pyzmq==22.3.0
requests==2.27.1
rsconnect-python==1.7.1
scikit-learn==1.0.2
scipy==1.8.0
semver==2.13.0
six==1.16.0
sklearn==0.0
sniffio==1.2.0
SQLAlchemy==1.4.36
stack-data==0.2.0
starlette==0.18.0
threadpoolctl==3.1.0
torch==1.11.0
tornado==6.1
traitlets==5.1.1
typing_extensions==4.2.0
urllib3==1.26.9
uvicorn==0.17.6
vetiver==0.1.3
wcwidth==0.2.5
xgboost==1.6.1
xxhash==3.0.0
zipp==3.8.0
